requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
gunicorn==21.2.0
